refactor instruments
