+++ apollo-router/src/plugins/telemetry/config_new/instruments.rs
@@ -31,7 +31,13 @@ use super::graphql::FIELD_LENGTH;
 use super::graphql::GraphQLInstruments;
 use super::graphql::selectors::ListLength;
 use super::http_server::attributes::HttpServerAttributes;
+use super::router::instruments::RouterInstruments;
+use super::router::instruments::RouterInstrumentsConfig;
 use super::selectors::CacheKind;
+use super::subgraph::instruments::SubgraphInstruments;
+use super::subgraph::instruments::SubgraphInstrumentsConfig;
+use super::supergraph::instruments::SupergraphCustomInstruments;
+use super::supergraph::instruments::SupergraphInstrumentsConfig;
 use crate::Context;
 use crate::axum_factory::connection_handle::ConnectionState;
 use crate::axum_factory::connection_handle::OPEN_CONNECTIONS_METRIC;
@@ -46,7 +52,6 @@ use crate::plugins::telemetry::config_new::connector::instruments::ConnectorInst
 use crate::plugins::telemetry::config_new::connector::selectors::ConnectorSelector;
 use crate::plugins::telemetry::config_new::connector::selectors::ConnectorValue;
 use crate::plugins::telemetry::config_new::cost::CostInstruments;
-use crate::plugins::telemetry::config_new::cost::CostInstrumentsConfig;
 use crate::plugins::telemetry::config_new::extendable::Extendable;
 use crate::plugins::telemetry::config_new::graphql::GraphQLInstrumentsConfig;
 use crate::plugins::telemetry::config_new::graphql::attributes::GraphQLAttributes;
@@ -65,7 +70,6 @@ use crate::plugins::telemetry::otlp::TelemetryDataKind;
 use crate::services::router;
 use crate::services::router::pipeline_handle::PIPELINE_METRIC;
 use crate::services::router::pipeline_handle::pipeline_counts;
-use crate::services::subgraph;
 use crate::services::supergraph;
 
 pub(crate) const METER_NAME: &str = "apollo/router";
@@ -1075,46 +1079,6 @@ impl StaticInstrument {
 
 #[derive(Clone, Deserialize, JsonSchema, Debug, Default)]
 #[serde(deny_unknown_fields, default)]
-pub(crate) struct RouterInstrumentsConfig {
-    /// Histogram of server request duration
-    #[serde(rename = "http.server.request.duration")]
-    http_server_request_duration:
-        DefaultedStandardInstrument<Extendable<RouterAttributes, RouterSelector>>,
-
-    /// Counter of active requests
-    #[serde(rename = "http.server.active_requests")]
-    http_server_active_requests: DefaultedStandardInstrument<ActiveRequestsAttributes>,
-
-    /// Histogram of server request body size
-    #[serde(rename = "http.server.request.body.size")]
-    http_server_request_body_size:
-        DefaultedStandardInstrument<Extendable<RouterAttributes, RouterSelector>>,
-
-    /// Histogram of server response body size
-    #[serde(rename = "http.server.response.body.size")]
-    http_server_response_body_size:
-        DefaultedStandardInstrument<Extendable<RouterAttributes, RouterSelector>>,
-}
-
-impl DefaultForLevel for RouterInstrumentsConfig {
-    fn defaults_for_level(
-        &mut self,
-        requirement_level: DefaultAttributeRequirementLevel,
-        kind: TelemetryDataKind,
-    ) {
-        self.http_server_request_duration
-            .defaults_for_levels(requirement_level, kind);
-        self.http_server_active_requests
-            .defaults_for_levels(requirement_level, kind);
-        self.http_server_request_body_size
-            .defaults_for_levels(requirement_level, kind);
-        self.http_server_response_body_size
-            .defaults_for_levels(requirement_level, kind);
-    }
-}
-
-#[derive(Clone, Deserialize, JsonSchema, Debug, Default)]
-#[serde(deny_unknown_fields, default)]
 pub(crate) struct ActiveRequestsAttributes {
     /// The HTTP request method
     #[serde(rename = "http.request.method")]
@@ -1239,56 +1203,6 @@ where
             Self::Bool(_) | Self::Unset => Vec::with_capacity(0),
             Self::Extendable { attributes } => attributes.on_response_event(response, ctx),
         }
-    }
-}
-
-#[derive(Clone, Deserialize, JsonSchema, Debug, Default)]
-#[serde(deny_unknown_fields, default)]
-pub(crate) struct SupergraphInstrumentsConfig {
-    #[serde(flatten)]
-    pub(crate) cost: CostInstrumentsConfig,
-}
-
-impl DefaultForLevel for SupergraphInstrumentsConfig {
-    fn defaults_for_level(
-        &mut self,
-        _requirement_level: DefaultAttributeRequirementLevel,
-        _kind: TelemetryDataKind,
-    ) {
-    }
-}
-
-#[derive(Clone, Deserialize, JsonSchema, Debug, Default)]
-#[serde(deny_unknown_fields, default)]
-pub(crate) struct SubgraphInstrumentsConfig {
-    /// Histogram of client request duration
-    #[serde(rename = "http.client.request.duration")]
-    http_client_request_duration:
-        DefaultedStandardInstrument<Extendable<SubgraphAttributes, SubgraphSelector>>,
-
-    /// Histogram of client request body size
-    #[serde(rename = "http.client.request.body.size")]
-    http_client_request_body_size:
-        DefaultedStandardInstrument<Extendable<SubgraphAttributes, SubgraphSelector>>,
-
-    /// Histogram of client response body size
-    #[serde(rename = "http.client.response.body.size")]
-    http_client_response_body_size:
-        DefaultedStandardInstrument<Extendable<SubgraphAttributes, SubgraphSelector>>,
-}
-
-impl DefaultForLevel for SubgraphInstrumentsConfig {
-    fn defaults_for_level(
-        &mut self,
-        requirement_level: DefaultAttributeRequirementLevel,
-        kind: TelemetryDataKind,
-    ) {
-        self.http_client_request_duration
-            .defaults_for_level(requirement_level, kind);
-        self.http_client_request_body_size
-            .defaults_for_level(requirement_level, kind);
-        self.http_client_response_body_size
-            .defaults_for_level(requirement_level, kind);
     }
 }
 
@@ -1467,32 +1381,6 @@ where
 
     fn on_error(&self, error: &BoxError, ctx: &Context) {
         self.attributes.on_error(error, ctx);
-    }
-}
-
-impl Selectors<subgraph::Request, subgraph::Response, ()> for SubgraphInstrumentsConfig {
-    fn on_request(&self, request: &subgraph::Request) -> Vec<opentelemetry::KeyValue> {
-        let mut attrs = self.http_client_request_body_size.on_request(request);
-        attrs.extend(self.http_client_request_duration.on_request(request));
-        attrs.extend(self.http_client_response_body_size.on_request(request));
-
-        attrs
-    }
-
-    fn on_response(&self, response: &subgraph::Response) -> Vec<opentelemetry::KeyValue> {
-        let mut attrs = self.http_client_request_body_size.on_response(response);
-        attrs.extend(self.http_client_request_duration.on_response(response));
-        attrs.extend(self.http_client_response_body_size.on_response(response));
-
-        attrs
-    }
-
-    fn on_error(&self, error: &BoxError, ctx: &Context) -> Vec<opentelemetry::KeyValue> {
-        let mut attrs = self.http_client_request_body_size.on_error(error, ctx);
-        attrs.extend(self.http_client_request_duration.on_error(error, ctx));
-        attrs.extend(self.http_client_response_body_size.on_error(error, ctx));
-
-        attrs
     }
 }
 
@@ -1726,74 +1614,6 @@ where
     }
 }
 
-pub(crate) struct RouterInstruments {
-    http_server_request_duration: Option<
-        CustomHistogram<router::Request, router::Response, (), RouterAttributes, RouterSelector>,
-    >,
-    http_server_active_requests: Option<ActiveRequestsCounter>,
-    http_server_request_body_size: Option<
-        CustomHistogram<router::Request, router::Response, (), RouterAttributes, RouterSelector>,
-    >,
-    http_server_response_body_size: Option<
-        CustomHistogram<router::Request, router::Response, (), RouterAttributes, RouterSelector>,
-    >,
-    custom: RouterCustomInstruments,
-}
-
-impl Instrumented for RouterInstruments {
-    type Request = router::Request;
-    type Response = router::Response;
-    type EventResponse = ();
-
-    fn on_request(&self, request: &Self::Request) {
-        if let Some(http_server_request_duration) = &self.http_server_request_duration {
-            http_server_request_duration.on_request(request);
-        }
-        if let Some(http_server_active_requests) = &self.http_server_active_requests {
-            http_server_active_requests.on_request(request);
-        }
-        if let Some(http_server_request_body_size) = &self.http_server_request_body_size {
-            http_server_request_body_size.on_request(request);
-        }
-        if let Some(http_server_response_body_size) = &self.http_server_response_body_size {
-            http_server_response_body_size.on_request(request);
-        }
-        self.custom.on_request(request);
-    }
-
-    fn on_response(&self, response: &Self::Response) {
-        if let Some(http_server_request_duration) = &self.http_server_request_duration {
-            http_server_request_duration.on_response(response);
-        }
-        if let Some(http_server_active_requests) = &self.http_server_active_requests {
-            http_server_active_requests.on_response(response);
-        }
-        if let Some(http_server_request_body_size) = &self.http_server_request_body_size {
-            http_server_request_body_size.on_response(response);
-        }
-        if let Some(http_server_response_body_size) = &self.http_server_response_body_size {
-            http_server_response_body_size.on_response(response);
-        }
-        self.custom.on_response(response);
-    }
-
-    fn on_error(&self, error: &BoxError, ctx: &Context) {
-        if let Some(http_server_request_duration) = &self.http_server_request_duration {
-            http_server_request_duration.on_error(error, ctx);
-        }
-        if let Some(http_server_active_requests) = &self.http_server_active_requests {
-            http_server_active_requests.on_error(error, ctx);
-        }
-        if let Some(http_server_request_body_size) = &self.http_server_request_body_size {
-            http_server_request_body_size.on_error(error, ctx);
-        }
-        if let Some(http_server_response_body_size) = &self.http_server_response_body_size {
-            http_server_response_body_size.on_error(error, ctx);
-        }
-        self.custom.on_error(error, ctx);
-    }
-}
-
 pub(crate) struct SupergraphInstruments {
     cost: CostInstruments,
     custom: SupergraphCustomInstruments,
@@ -1824,109 +1644,6 @@ impl Instrumented for SupergraphInstruments {
         self.custom.on_response_event(response, ctx);
     }
 }
-
-pub(crate) struct SubgraphInstruments {
-    http_client_request_duration: Option<
-        CustomHistogram<
-            subgraph::Request,
-            subgraph::Response,
-            (),
-            SubgraphAttributes,
-            SubgraphSelector,
-        >,
-    >,
-    http_client_request_body_size: Option<
-        CustomHistogram<
-            subgraph::Request,
-            subgraph::Response,
-            (),
-            SubgraphAttributes,
-            SubgraphSelector,
-        >,
-    >,
-    http_client_response_body_size: Option<
-        CustomHistogram<
-            subgraph::Request,
-            subgraph::Response,
-            (),
-            SubgraphAttributes,
-            SubgraphSelector,
-        >,
-    >,
-    custom: SubgraphCustomInstruments,
-}
-
-impl Instrumented for SubgraphInstruments {
-    type Request = subgraph::Request;
-    type Response = subgraph::Response;
-    type EventResponse = ();
-
-    fn on_request(&self, request: &Self::Request) {
-        if let Some(http_client_request_duration) = &self.http_client_request_duration {
-            http_client_request_duration.on_request(request);
-        }
-        if let Some(http_client_request_body_size) = &self.http_client_request_body_size {
-            http_client_request_body_size.on_request(request);
-        }
-        if let Some(http_client_response_body_size) = &self.http_client_response_body_size {
-            http_client_response_body_size.on_request(request);
-        }
-        self.custom.on_request(request);
-    }
-
-    fn on_response(&self, response: &Self::Response) {
-        if let Some(http_client_request_duration) = &self.http_client_request_duration {
-            http_client_request_duration.on_response(response);
-        }
-        if let Some(http_client_request_body_size) = &self.http_client_request_body_size {
-            http_client_request_body_size.on_response(response);
-        }
-        if let Some(http_client_response_body_size) = &self.http_client_response_body_size {
-            http_client_response_body_size.on_response(response);
-        }
-        self.custom.on_response(response);
-    }
-
-    fn on_error(&self, error: &BoxError, ctx: &Context) {
-        if let Some(http_client_request_duration) = &self.http_client_request_duration {
-            http_client_request_duration.on_error(error, ctx);
-        }
-        if let Some(http_client_request_body_size) = &self.http_client_request_body_size {
-            http_client_request_body_size.on_error(error, ctx);
-        }
-        if let Some(http_client_response_body_size) = &self.http_client_response_body_size {
-            http_client_response_body_size.on_error(error, ctx);
-        }
-        self.custom.on_error(error, ctx);
-    }
-}
-
-pub(crate) type RouterCustomInstruments = CustomInstruments<
-    router::Request,
-    router::Response,
-    (),
-    RouterAttributes,
-    RouterSelector,
-    RouterValue,
->;
-
-pub(crate) type SupergraphCustomInstruments = CustomInstruments<
-    supergraph::Request,
-    supergraph::Response,
-    crate::graphql::Response,
-    SupergraphAttributes,
-    SupergraphSelector,
-    SupergraphValue,
->;
-
-pub(crate) type SubgraphCustomInstruments = CustomInstruments<
-    subgraph::Request,
-    subgraph::Response,
-    (),
-    SubgraphAttributes,
-    SubgraphSelector,
-    SubgraphValue,
->;
 
 // ---------------- Counter -----------------------
 #[derive(Debug, Clone)]
@@ -2313,7 +2030,7 @@ where
     }
 }
 
-struct ActiveRequestsCounter {
+pub(crate) struct ActiveRequestsCounter {
     inner: Mutex<ActiveRequestsCounterInner>,
 }
 
@@ -2768,6 +2485,7 @@ mod tests {
     use crate::plugins::telemetry::config_new::graphql::GraphQLInstruments;
     use crate::plugins::telemetry::config_new::instruments::Instrumented;
     use crate::plugins::telemetry::config_new::instruments::InstrumentsConfig;
+    use crate::plugins::telemetry::config_new::supergraph::instruments::SupergraphCustomInstruments;
     use crate::services::OperationKind;
     use crate::services::RouterRequest;
     use crate::services::RouterResponse;
@@ -3219,7 +2937,7 @@ mod tests {
                                     let mut http_request = http::Request::new(graphql_request);
                                     *http_request.headers_mut() = convert_http_headers(headers);
 
-                                    let request = subgraph::Request::fake_builder()
+                                    let request = crate::plugins::telemetry::subgraph::Request::fake_builder()
                                         .context(context.clone())
                                         .subgraph_name(subgraph_name)
                                         .and_operation_kind(operation_kind)
@@ -3237,7 +2955,7 @@ mod tests {
                                     errors,
                                     headers,
                                 } => {
-                                    let response = subgraph::Response::fake2_builder()
+                                    let response = crate::plugins::telemetry::subgraph::Response::fake2_builder()
                                         .context(context.clone())
                                         .and_subgraph_name(subgraph_name)
                                         .status_code(StatusCode::from_u16(status).expect("status"))
