+++ apollo-router/src/plugins/telemetry/mod.rs
@@ -65,11 +65,11 @@ use self::apollo_exporter::proto;
 use self::config::Conf;
 use self::config::TraceIdFormat;
 use self::config_new::instruments::Instrumented;
-use self::config_new::instruments::RouterInstruments;
-use self::config_new::instruments::SubgraphInstruments;
 use self::config_new::router::events::RouterEvents;
+use self::config_new::router::instruments::RouterInstruments;
 use self::config_new::spans::Spans;
 use self::config_new::subgraph::events::SubgraphEvents;
+use self::config_new::subgraph::instruments::SubgraphInstruments;
 use self::config_new::supergraph::events::SupergraphEvents;
 use self::metrics::apollo::studio::SingleTypeStat;
 use self::reload::reload_fmt;
