+++ apollo-router/src/services/router/service.rs
@@ -61,10 +61,10 @@ use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_BODY;
 use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_HEADERS;
 use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_URI;
 use crate::plugins::telemetry::config_new::attributes::HTTP_REQUEST_VERSION;
-use crate::plugins::telemetry::config_new::events::DisplayRouterRequest;
-use crate::plugins::telemetry::config_new::events::DisplayRouterResponse;
-use crate::plugins::telemetry::config_new::events::RouterResponseBodyExtensionType;
 use crate::plugins::telemetry::config_new::events::log_event;
+use crate::plugins::telemetry::config_new::router::events::DisplayRouterRequest;
+use crate::plugins::telemetry::config_new::router::events::DisplayRouterResponse;
+use crate::plugins::telemetry::config_new::router::events::RouterResponseBodyExtensionType;
 use crate::protocols::multipart::Multipart;
 use crate::protocols::multipart::ProtocolMode;
 use crate::query_planner::InMemoryCachePlanner;
